requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.2.2
click==8.1.7
fastapi==0.110.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
magic-filter==1.0.12
multidict==6.0.5
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
typing_extensions==4.10.0
tzlocal==5.2
uvicorn==0.28.0
yarl==1.9.4
